fix portability problem between 32 and 64 bit as long is not a portable type. -long-This line, and those below, will be ignored--
M    TMySQLStatement.cxx


git-svn-id: http://root.cern.ch/svn/root/branches/v5-22-00-patches@29277 27541ba8-7e3a-0410-8455-c3a389f83636

+++ sql/mysql/src/TMySQLStatement.cxx
@@ -335,8 +335,8 @@ const char* TMySQLStatement::ConvertToString(Int_t npar)
 
    switch(fBind[npar].buffer_type) {
       case MYSQL_TYPE_LONG:
-         if (sig) snprintf(buf,100,"%ld",*((long*) addr));
-             else snprintf(buf,100,"%lu",*((unsigned long*) addr));
+         if (sig) snprintf(buf,100,"%d",*((int*) addr));
+             else snprintf(buf,100,"%u",*((unsigned int*) addr));
          break;
       case MYSQL_TYPE_LONGLONG:
          if (sig) snprintf(buf,100,"%lld",*((long long*) addr)); else
